Moved notification plugin management to notification_service.dart; Added file headers to new dart files
