+++ lib/home_page.dart
@@ -10,6 +10,7 @@
 import 'dart:async';
 import 'package:app_screen_time/main.dart';
 import 'package:app_screen_time/notifications.dart';
+import 'package:app_screen_time/notification_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:intl/intl.dart';
