+++ lib/main.dart
@@ -9,6 +9,7 @@
 //Dart Imports
 import 'dart:async';
 import 'dart:io';
+import 'package:app_screen_time/notification_service.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -25,8 +26,6 @@ import 'home_page.dart';
 import 'social_media_page.dart';
 import 'historical_data_page.dart';
 import 'login_screen.dart';
-
-final FlutterLocalNotificationsPlugin notificationsPlugin = FlutterLocalNotificationsPlugin();
 
 ///*********************************
 /// Name: main
@@ -86,46 +85,12 @@ class _MyPageViewState extends State<MyPageView> {
   //Tracks current index
   int _currentPage = 0;
 
-  bool _notificationsEnabled = false;
-
   //Initialize page controller and set initial page
   @override
   void initState() {
     _pageController = PageController(initialPage: 1);
     _currentPage = 1;
-
-    const InitializationSettings initSettings = InitializationSettings(
-      android: AndroidInitializationSettings("@mipmap/ic_launcher")
-    );
-
-    notificationsPlugin.initialize(
-      initSettings
-    );
-
-    Future<void> _isAndroidPermissionGranted() async {
-      final bool granted = await notificationsPlugin
-              .resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()
-              ?.areNotificationsEnabled() ??
-          false;
-
-      setState(() {
-        _notificationsEnabled = granted;
-      });
-    }
-
-    Future<void> _requestPermissions() async {
-      final AndroidFlutterLocalNotificationsPlugin? androidImplementation = 
-        notificationsPlugin.resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>();
-
-      final bool? grantedNotificationPermission =
-        await androidImplementation?.requestNotificationsPermission();
-      setState(() {
-        _notificationsEnabled = grantedNotificationPermission ?? false;
-      });
-    }
-
-    _isAndroidPermissionGranted();
-    _requestPermissions();
+    NotificationService.initialize();
     super.initState();
   }
